ZC-9550: remove bit we don’t need after all
looking closer I think this was a one time fix thing due to how package
manager worked. it basically says "if an app’s ruby was not explicitly
set set it to the original ruby24 since that is what they are using" so
yeah we don’t need to do that after all, hurray! and also a good
reminder why we need to drop this mess lol